Rmv invoking parens on callback
In two places you wrap an anonymous function in `$()`, for the stated
purpose of ensuring that the code in the function waits until browser
ready before executing, however you invoke the anonymous function with
parentheses.

This means that the code in the function will immediately execute, and
the return value of the function will be given to `$()`. Since the
return value is not a function, this contradicts the correct usage of
the jQuery method, and contradicts the stated purpose of using $().